Update SensorSample sensor values to Double

+++ app/src/main/java/com/example/noimo/domain/SensorSample.kt
@@ -3,6 +3,6 @@ package com.example.noimo.domain
 // Holds sensor data
 data class SensorSample (
     val timestampMillis: Long,
-    val accelerationMagnitude: Float,
-    val audioAmplitude: Float
+    val accelerationMagnitude: Double,
+    val audioAmplitude: Double
 )
